# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Zora Sniper API - Auto trade bot for Zora Coins on Base
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface BotStatus {
  running: boolean;
  /** @nullable */
  walletAddress: string | null;
  /** @nullable */
  walletBalanceEth: string | null;
  totalTrades: number;
  snipedToday?: number;
  uptimeSeconds: number;
  /** @nullable */
  lastEventAt?: string | null;
  network?: string;
}

/**
 * Whitelist = only listed creators, all = any new coin
 */
export type BotConfigWatchMode = typeof BotConfigWatchMode[keyof typeof BotConfigWatchMode];


export const BotConfigWatchMode = {
  whitelist: 'whitelist',
  all: 'all',
} as const;

export interface BotConfig {
  /** ETH amount per snipe e.g. "0.01" */
  buyAmountEth: string;
  /** Slippage tolerance in percent */
  slippagePercent: number;
  /** Max gas price in Gwei */
  maxGasGwei: number;
  /** Whitelist = only listed creators, all = any new coin */
  watchMode: BotConfigWatchMode;
  enabled: boolean;
  /** @nullable */
  minLiquidityEth?: number | null;
  autoSell?: boolean;
  /** @nullable */
  takeProfitPercent?: number | null;
  /** @nullable */
  stopLossPercent?: number | null;
}

export type ConfigUpdateWatchMode = typeof ConfigUpdateWatchMode[keyof typeof ConfigUpdateWatchMode];


export const ConfigUpdateWatchMode = {
  whitelist: 'whitelist',
  all: 'all',
} as const;

export interface ConfigUpdate {
  buyAmountEth?: string;
  slippagePercent?: number;
  maxGasGwei?: number;
  watchMode?: ConfigUpdateWatchMode;
  enabled?: boolean;
  /** @nullable */
  minLiquidityEth?: number | null;
  autoSell?: boolean;
  /** @nullable */
  takeProfitPercent?: number | null;
  /** @nullable */
  stopLossPercent?: number | null;
}

export interface Creator {
  address: string;
  label: string;
  enabled: boolean;
  addedAt: string;
  totalSniped: number;
  /** @nullable */
  zoraProfileUrl?: string | null;
}

export interface CreatorInput {
  /** Ethereum address of the creator */
  address: string;
  /** Optional display name/label */
  label?: string;
}

export interface CreatorPatch {
  label?: string;
  enabled?: boolean;
}

export type TradeStatus = typeof TradeStatus[keyof typeof TradeStatus];


export const TradeStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  failed: 'failed',
  sold: 'sold',
} as const;

export interface Trade {
  id: number;
  /** @nullable */
  txHash?: string | null;
  tokenAddress: string;
  tokenName: string;
  tokenSymbol: string;
  creatorAddress: string;
  buyAmountEth: string;
  /** @nullable */
  tokenAmount?: string | null;
  /** @nullable */
  gasUsedEth?: string | null;
  timestamp: string;
  status: TradeStatus;
  /** @nullable */
  sellTxHash?: string | null;
  /** @nullable */
  sellAmountEth?: string | null;
  /** @nullable */
  pnlEth?: string | null;
  /** @nullable */
  blockNumber?: number | null;
}

export interface TradeStats {
  totalTrades: number;
  successfulTrades: number;
  failedTrades: number;
  totalEthSpent: string;
  totalEthRecovered: string;
  totalPnlEth: string;
  winCount: number;
  lossCount: number;
  winRatePercent?: number;
  avgBuyAmountEth?: string;
  todayTrades?: number;
  todayEthSpent?: string;
}

export interface DashboardData {
  botStatus: BotStatus;
  recentTrades: Trade[];
  stats: TradeStats;
  topCreators: Creator[];
}

export type ListTradesParams = {
limit?: number;
offset?: number;
status?: ListTradesStatus;
};

export type ListTradesStatus = typeof ListTradesStatus[keyof typeof ListTradesStatus];


export const ListTradesStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  failed: 'failed',
  sold: 'sold',
} as const;