# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Zora Sniper API - Auto trade bot for Zora Coins on Base
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /creators:
    get:
      operationId: listCreators
      tags: [creators]
      summary: List whitelisted creators
      responses:
        "200":
          description: List of creators
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Creator"
    post:
      operationId: addCreator
      tags: [creators]
      summary: Add creator to whitelist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatorInput"
      responses:
        "201":
          description: Creator added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Creator"
        "400":
          description: Invalid address or already exists
  /creators/{address}:
    delete:
      operationId: removeCreator
      tags: [creators]
      summary: Remove creator from whitelist
      parameters:
        - name: address
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Removed
        "404":
          description: Not found
    patch:
      operationId: updateCreator
      tags: [creators]
      summary: Update creator label or enabled state
      parameters:
        - name: address
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatorPatch"
      responses:
        "200":
          description: Updated creator
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Creator"
...
components:
  schemas:
...
    Creator:
      type: object
      required: [address, label, enabled, addedAt, totalSniped]
      properties:
        address:
          type: string
        label:
          type: string
        enabled:
          type: boolean
        addedAt:
          type: string
          format: date-time
        totalSniped:
          type: integer
        zoraProfileUrl:
          type: ["string", "null"]
    CreatorInput:
      type: object
      required: [address]
      properties:
        address:
          type: string
          description: Ethereum address of the creator
        label:
          type: string
          description: Optional display name/label
    CreatorPatch:
      type: object
      properties:
        label:
          type: string
        enabled:
          type: boolean
...