# lib/api-zod/src/generated/types/creatorPatch.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Zora Sniper API - Auto trade bot for Zora Coins on Base
 * OpenAPI spec version: 0.1.0
 */

export interface CreatorPatch {
  label?: string;
  enabled?: boolean;
}